errors removed

+++ Online_Shopping_Application_Backend/src/main/java/com/masai/repository/OrderRepository.java
@@ -1,15 +1,14 @@
 package com.masai.repository;
 
 import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-
 import com.masai.model.Order;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+	
 	@Query("SELECT o FROM Order o WHERE o.orderStatus = :status")
     List<Order> findOrdersByStatus(String status);
